allow run with no schedules, adjust warning
