+++ core/src/tasks/system/status/status.ts
@@ -1,4 +1,5 @@
 import { api, log, task } from "actionhero";
+import { Schedule } from "../../../models/Schedule";
 import { GrouparooCLI } from "../../../modules/cli";
 import { APM } from "../../../modules/apm";
 import { Status, FinalSummary } from "../../../modules/status";
@@ -38,6 +39,7 @@ export class StatusTask extends CLSTask {
       const complete = await this.checkForComplete(samples);
       if (runMode === "cli:run" && complete) {
         await this.logFinalSummary();
+        await this.checkSchedules();
         await this.stopServer(toStop);
       }
 
@@ -48,6 +50,18 @@ export class StatusTask extends CLSTask {
   async logFinalSummary() {
     const finalSummaryLog = await FinalSummary.getFinalSummary();
     GrouparooCLI.logger.finalSummary(finalSummaryLog);
+  }
+
+  async checkSchedules() {
+    const scheduleCount = await Schedule.count();
+    if (scheduleCount === 0) {
+      return log(
+        `No schedules found.
+The run command uses schedules to know what profiles to import.
+See this link for more info: https://www.grouparoo.com/docs/getting-started/product-concepts#schedule`,
+        "warning"
+      );
+    }
   }
 
   async checkForComplete(samples: Status.StatusGetResponse) {
